agents/budget_agent/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
email-validator
openai
google-generativeai
faiss-cpu
scikit-learn
pandas
numpy
python-multipart
python-dotenv
statsmodels
requests
python-jose[cryptography]
passlib
apscheduler
pytest
alembic
redis
langgraph
langgraph-checkpoint-sqlite
